requirements.txt
# external requirements
click
Sphinx
coverage
awscli>=1.27.130
python-dotenv>=0.5.1
flake8
requests
requests-mock
scipy
pandas
shapely
rasterio
xarray
pyproj
matplotlib
fastapi
hypercorn[trio]==0.14.3
zarr
geojson
numpy
geopandas
pystac-client
rioxarray
h5netcdf
netcdf4
pystac
stackstac
odc-stac
earthdaily==0.0.4
# aws
boto3
# azure
azure-storage-blob
byoa